reduce code in MainActivity
