+++ app/src/main/java/com/aneke/peter/ctwnews/MainActivity.kt
@@ -3,16 +3,13 @@ package com.aneke.peter.ctwnews
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.widget.Toast
-import androidx.biometric.BiometricManager
 import androidx.biometric.BiometricPrompt
-import androidx.core.content.ContextCompat
 import com.aneke.peter.ctwnews.news.NewsActivity
-import java.util.concurrent.Executor
+import com.aneke.peter.ctwnews.utils.deviceHasBiometric
+import com.aneke.peter.ctwnews.utils.setupBiometricPromptInfo
 
 class MainActivity : AppCompatActivity() {
 
-    private lateinit var executor: Executor
     private lateinit var biometricPrompt: BiometricPrompt
     private lateinit var promptInfo: BiometricPrompt.PromptInfo
 
@@ -20,63 +17,22 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
-        executor = ContextCompat.getMainExecutor(this)
-        biometricPrompt = BiometricPrompt(this, executor,
-            object : BiometricPrompt.AuthenticationCallback() {
-                override fun onAuthenticationError(errorCode: Int,
-                                                   errString: CharSequence) {
-                    super.onAuthenticationError(errorCode, errString)
-                    Toast.makeText(applicationContext,
-                        "Authentication error: $errString", Toast.LENGTH_SHORT)
-                        .show()
-                }
-
-                override fun onAuthenticationSucceeded(
-                    result: BiometricPrompt.AuthenticationResult) {
-                    super.onAuthenticationSucceeded(result)
-                    Toast.makeText(applicationContext,
-                        "Authentication succeeded!", Toast.LENGTH_SHORT)
-                        .show()
-                    showHeadlines()
-                }
-
-                override fun onAuthenticationFailed() {
-                    super.onAuthenticationFailed()
-                    Toast.makeText(applicationContext, "Authentication failed",
-                        Toast.LENGTH_SHORT)
-                        .show()
-                   finishAffinity()
-                }
-            })
-
-        promptInfo = BiometricPrompt.PromptInfo.Builder()
-            .setTitle("Biometric login for News App")
-            .setNegativeButtonText("Cancel")
-            .build()
-
+        setupPrompt()
         checkDeviceHasBiometric()
     }
 
+    private fun setupPrompt() {
+        setupBiometricPromptInfo({ _biometricPrompt, _promptInfo ->
+            biometricPrompt = _biometricPrompt
+            promptInfo = _promptInfo
+        }, {showHeadlines()})
+    }
+
     private fun checkDeviceHasBiometric() {
-        val biometricManager = BiometricManager.from(this)
-        when (biometricManager.canAuthenticate(BiometricManager.Authenticators.BIOMETRIC_STRONG
-                or BiometricManager.Authenticators.DEVICE_CREDENTIAL)) {
-            BiometricManager.BIOMETRIC_SUCCESS -> {
-                promptBiometrics()
-            }
-            BiometricManager.BIOMETRIC_ERROR_NO_HARDWARE -> {
-                showHeadlines()
-            }
-            BiometricManager.BIOMETRIC_ERROR_HW_UNAVAILABLE -> {
-                showHeadlines()
-            }
-            BiometricManager.BIOMETRIC_ERROR_NONE_ENROLLED -> {
-                showHeadlines()
-            }
-            BiometricManager.BIOMETRIC_ERROR_SECURITY_UPDATE_REQUIRED,
-            BiometricManager.BIOMETRIC_ERROR_UNSUPPORTED, BiometricManager.BIOMETRIC_STATUS_UNKNOWN -> {
-                showHeadlines()
-            }
+        if (deviceHasBiometric()) {
+            promptBiometrics()
+        }else {
+            showHeadlines()
         }
     }
