+++ app/src/main/java/com/aneke/peter/ctwnews/utils/Extensions.kt
@@ -4,7 +4,12 @@ import android.app.Activity
 import android.content.Intent
 import android.os.Build.VERSION.SDK_INT
 import android.os.Parcelable
+import android.widget.Toast
 import androidx.appcompat.app.AlertDialog
+import androidx.biometric.BiometricManager
+import androidx.biometric.BiometricPrompt
+import androidx.core.content.ContextCompat
+import androidx.fragment.app.FragmentActivity
 import com.aneke.peter.ctwnews.BuildConfig
 import org.json.JSONException
 import org.json.JSONObject
@@ -76,4 +81,71 @@ fun String.convertToDate() : String {
 inline fun <reified T : Parcelable> Intent.parcelable(key: String): T? = when {
     SDK_INT >= 33 -> getParcelableExtra(key, T::class.java)
     else -> @Suppress("DEPRECATION") getParcelableExtra(key) as? T
+}
+
+fun Activity.deviceHasBiometric() : Boolean {
+    val biometricManager = BiometricManager.from(this)
+    return when (biometricManager.canAuthenticate(
+        BiometricManager.Authenticators.BIOMETRIC_STRONG
+            or BiometricManager.Authenticators.DEVICE_CREDENTIAL)) {
+        BiometricManager.BIOMETRIC_SUCCESS -> {
+            true
+        }
+        BiometricManager.BIOMETRIC_ERROR_NO_HARDWARE -> {
+            false
+        }
+        BiometricManager.BIOMETRIC_ERROR_HW_UNAVAILABLE -> {
+            false
+        }
+        BiometricManager.BIOMETRIC_ERROR_NONE_ENROLLED -> {
+            false
+        }
+        BiometricManager.BIOMETRIC_ERROR_SECURITY_UPDATE_REQUIRED,
+        BiometricManager.BIOMETRIC_ERROR_UNSUPPORTED, BiometricManager.BIOMETRIC_STATUS_UNKNOWN -> {
+            false
+        }
+        else -> {
+            false
+        }
+    }
+}
+
+fun FragmentActivity.setupBiometricPromptInfo(assign : (BiometricPrompt, BiometricPrompt.PromptInfo) -> Unit, success : () -> Unit) {
+    val executor = ContextCompat.getMainExecutor(this)
+
+    val biometricPrompt = BiometricPrompt(this, executor,
+        object : BiometricPrompt.AuthenticationCallback() {
+            override fun onAuthenticationError(errorCode: Int,
+                                               errString: CharSequence) {
+                super.onAuthenticationError(errorCode, errString)
+                Toast.makeText(applicationContext,
+                    "Authentication error: $errString", Toast.LENGTH_SHORT)
+                    .show()
+            }
+
+            override fun onAuthenticationSucceeded(
+                result: BiometricPrompt.AuthenticationResult) {
+                super.onAuthenticationSucceeded(result)
+                Toast.makeText(applicationContext,
+                    "Authentication succeeded!", Toast.LENGTH_SHORT)
+                    .show()
+                success()
+            }
+
+            override fun onAuthenticationFailed() {
+                super.onAuthenticationFailed()
+                Toast.makeText(applicationContext, "Authentication failed",
+                    Toast.LENGTH_SHORT)
+                    .show()
+                finishAffinity()
+            }
+        })
+
+    val promptInfo = BiometricPrompt.PromptInfo.Builder()
+        .setTitle("Biometric login for News App")
+        .setNegativeButtonText("Cancel")
+        .build()
+
+    assign(biometricPrompt, promptInfo)
+
 }
